backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0
aiosqlite==0.19.0
# # WaveRider Production Backend Dependencies

# Core FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & Caching
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis[hiredis]==5.0.1
alembic==1.13.0

# AI Services
openai==1.3.5
anthropic==0.7.8
httpx==0.25.2

# Vector Database
pinecone-client==2.2.4

# File Operations
aiofiles==23.2.1

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation & Serialization
pydantic==2.5.0

# Environment & Configuration
python-dotenv==1.0.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# WebSocket Support
websockets==12.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Production WSGI
gunicorn==21.2.0
celery==5.3.4
websockets==12.0
socketio==0.2.1
python-socketio==5.10.0
aiofiles==23.2.1
# aioredis==2.0.1
httpx==0.25.2
openai==1.3.7
anthropic==0.7.8
langchain==0.0.350
langchain-community==0.0.1
langgraph==0.0.20
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
transformers==4.36.0
torch==2.1.1
sentence-transformers==2.2.2
pypdf2==3.0.1
python-magic==0.4.27
gitpython==3.1.40
docker==6.1.3
kubernetes==28.1.0
boto3==1.34.0
google-cloud-storage==2.10.0
azure-storage-blob==12.19.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
coverage==7.3.2
